@@ -6,30 +6,13 @@ import 'dart:ui' show hashValues;
 import 'package:flutter/foundation.dart';
 
 import 'types.dart';
-import 'package:meta/meta.dart' show immutable, required;
+import 'package:meta/meta.dart' show immutable;
 
 /// Uniquely identifies a [TileOverlay] among [GoogleMap] tile overlays.
 @immutable
-class TileOverlayId {
+class TileOverlayId extends MapsObjectId<TileOverlay> {
   /// Creates an immutable identifier for a [TileOverlay].
-  TileOverlayId(this.value) : assert(value != null);
-
-  /// The value of the [TileOverlayId].
-  final String value;
-
-  @override
-  bool operator ==(Object other) {
-    if (other.runtimeType != runtimeType) {
-      return false;
-    }
-    return other is TileOverlayId && other.value == value;
-  }
-
-  @override
-  int get hashCode => value.hashCode;
-
-  @override
-  String toString() => '${objectRuntimeType(this, 'TileOverlayId')}($value)';
+  TileOverlayId(String value) : super(value);
 }
 
 /// A set of images which are displayed on top of the base map tiles.
@@ -61,14 +44,14 @@ class TileOverlayId {
 /// At zoom level N, the x values of the tile coordinates range from 0 to 2N - 1 and increase from
 /// west to east and the y values range from 0 to 2N - 1 and increase from north to south.
 ///
-class TileOverlay {
+class TileOverlay implements MapsObject {
   /// Creates an immutable representation of a [TileOverlay] to draw on [GoogleMap].
   const TileOverlay({
-    @required this.tileOverlayId,
+    required this.tileOverlayId,
     this.fadeIn = true,
     this.tileProvider,
     this.transparency = 0.0,
-    this.zIndex,
+    this.zIndex = 0,
     this.visible = true,
     this.tileSize = 256,
   }) : assert(transparency >= 0.0 && transparency <= 1.0);
@@ -76,11 +59,14 @@ class TileOverlay {
   /// Uniquely identifies a [TileOverlay].
   final TileOverlayId tileOverlayId;
 
+  @override
+  TileOverlayId get mapsId => tileOverlayId;
+
   /// Whether the tiles should fade in. The default is true.
   final bool fadeIn;
 
   /// The tile provider to use for this tile overlay.
-  final TileProvider tileProvider;
+  final TileProvider? tileProvider;
 
   /// The transparency of the tile overlay. The default transparency is 0 (opaque).
   final double transparency;
@@ -104,12 +90,11 @@ class TileOverlay {
   /// Creates a new [TileOverlay] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
   TileOverlay copyWith({
-    TileOverlayId tileOverlayId,
-    bool fadeInParam,
-    double transparencyParam,
-    int zIndexParam,
-    bool visibleParam,
-    int tileSizeParam,
+    bool? fadeInParam,
+    double? transparencyParam,
+    int? zIndexParam,
+    bool? visibleParam,
+    int? tileSizeParam,
   }) {
     return TileOverlay(
       tileOverlayId: tileOverlayId,
@@ -121,11 +106,13 @@ class TileOverlay {
     );
   }
 
-  /// Converts this object to JSON.
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> json = <String, dynamic>{};
+  TileOverlay clone() => copyWith();
 
-    void addIfPresent(String fieldName, dynamic value) {
+  /// Converts this object to JSON.
+  Object toJson() {
+    final Map<String, Object> json = <String, Object>{};
+
+    void addIfPresent(String fieldName, Object? value) {
       if (value != null) {
         json[fieldName] = value;
       }
