@@ -10,14 +10,14 @@ class PatternItem {
   const PatternItem._(this._json);
 
   /// A dot used in the stroke pattern for a [Polyline].
-  static const PatternItem dot = PatternItem._(<dynamic>['dot']);
+  static const PatternItem dot = PatternItem._(<Object>['dot']);
 
   /// A dash used in the stroke pattern for a [Polyline].
   ///
   /// [length] has to be non-negative.
   static PatternItem dash(double length) {
     assert(length >= 0.0);
-    return PatternItem._(<dynamic>['dash', length]);
+    return PatternItem._(<Object>['dash', length]);
   }
 
   /// A gap used in the stroke pattern for a [Polyline].
@@ -25,11 +25,11 @@ class PatternItem {
   /// [length] has to be non-negative.
   static PatternItem gap(double length) {
     assert(length >= 0.0);
-    return PatternItem._(<dynamic>['gap', length]);
+    return PatternItem._(<Object>['gap', length]);
   }
 
-  final dynamic _json;
+  final Object _json;
 
   /// Converts this object to something serializable in JSON.
-  dynamic toJson() => _json;
+  Object toJson() => _json;
 }
