@@ -3,20 +3,14 @@
 // found in the LICENSE file.
 
 import '../types.dart';
+import 'maps_object.dart';
 
 /// Converts an [Iterable] of Markers in a Map of MarkerId -> Marker.
 Map<MarkerId, Marker> keyByMarkerId(Iterable<Marker> markers) {
-  if (markers == null) {
-    return <MarkerId, Marker>{};
-  }
-  return Map<MarkerId, Marker>.fromEntries(markers.map((Marker marker) =>
-      MapEntry<MarkerId, Marker>(marker.markerId, marker.clone())));
+  return keyByMapsObjectId<Marker>(markers).cast<MarkerId, Marker>();
 }
 
 /// Converts a Set of Markers into something serializable in JSON.
-List<Map<String, dynamic>> serializeMarkerSet(Set<Marker> markers) {
-  if (markers == null) {
-    return null;
-  }
-  return markers.map<Map<String, dynamic>>((Marker m) => m.toJson()).toList();
+Object serializeMarkerSet(Set<Marker> markers) {
+  return serializeMapsObjectSet(markers);
 }
