@@ -3,20 +3,14 @@
 // found in the LICENSE file.
 
 import '../types.dart';
+import 'maps_object.dart';
 
 /// Converts an [Iterable] of Polygons in a Map of PolygonId -> Polygon.
 Map<PolygonId, Polygon> keyByPolygonId(Iterable<Polygon> polygons) {
-  if (polygons == null) {
-    return <PolygonId, Polygon>{};
-  }
-  return Map<PolygonId, Polygon>.fromEntries(polygons.map((Polygon polygon) =>
-      MapEntry<PolygonId, Polygon>(polygon.polygonId, polygon.clone())));
+  return keyByMapsObjectId<Polygon>(polygons).cast<PolygonId, Polygon>();
 }
 
 /// Converts a Set of Polygons into something serializable in JSON.
-List<Map<String, dynamic>> serializePolygonSet(Set<Polygon> polygons) {
-  if (polygons == null) {
-    return null;
-  }
-  return polygons.map<Map<String, dynamic>>((Polygon p) => p.toJson()).toList();
+Object serializePolygonSet(Set<Polygon> polygons) {
+  return serializeMapsObjectSet(polygons);
 }
