@@ -1,23 +1,18 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import '../types.dart';
+import 'maps_object.dart';
 
 /// Converts an [Iterable] of TileOverlay in a Map of TileOverlayId -> TileOverlay.
 Map<TileOverlayId, TileOverlay> keyTileOverlayId(
     Iterable<TileOverlay> tileOverlays) {
-  if (tileOverlays == null) {
-    return <TileOverlayId, TileOverlay>{};
-  }
-  return Map<TileOverlayId, TileOverlay>.fromEntries(tileOverlays.map(
-      (TileOverlay tileOverlay) => MapEntry<TileOverlayId, TileOverlay>(
-          tileOverlay.tileOverlayId, tileOverlay)));
+  return keyByMapsObjectId<TileOverlay>(tileOverlays)
+      .cast<TileOverlayId, TileOverlay>();
 }
 
 /// Converts a Set of TileOverlays into something serializable in JSON.
-List<Map<String, dynamic>> serializeTileOverlaySet(
-    Set<TileOverlay> tileOverlays) {
-  if (tileOverlays == null) {
-    return null;
-  }
-  return tileOverlays
-      .map<Map<String, dynamic>>((TileOverlay p) => p.toJson())
-      .toList();
+Object serializeTileOverlaySet(Set<TileOverlay> tileOverlays) {
+  return serializeMapsObjectSet(tileOverlays);
 }
