@@ -13,16 +13,21 @@ void main() {
     test('toJson returns correct format', () async {
       final Uint8List data = Uint8List.fromList([0, 1]);
       final Tile tile = Tile(100, 200, data);
-      final Map<String, dynamic> json = tile.toJson();
-      expect(json['width'], 100);
-      expect(json['height'], 200);
-      expect(json['data'], data);
+      final Object json = tile.toJson();
+      expect(json, <String, Object>{
+        'width': 100,
+        'height': 200,
+        'data': data,
+      });
     });
 
-    test('toJson returns empty if nothing presents', () async {
-      final Tile tile = Tile(null, null, null);
-      final Map<String, dynamic> json = tile.toJson();
-      expect(json.isEmpty, true);
+    test('toJson handles null data', () async {
+      final Tile tile = Tile(0, 0, null);
+      final Object json = tile.toJson();
+      expect(json, <String, Object>{
+        'width': 0,
+        'height': 0,
+      });
     });
   });
 }
